build: added the ability to seed data based on env - improved env file for docker compose
