+++ .env.development
@@ -7,5 +7,4 @@ MONGO_REPLICA_PORT=27018
 
 # Server configurations
 DATABASE_URL=mongodb://admin:password@mongo_replica:27018/chronicle?authSource=admin
-# Could be DEV or STAGING or PROD
 CHRONICLE_ENV=DEV
